renamed stepdefs folder to check if the test are recognized
