--- src/test/java/com/example/Team1webshop/StepDefs/CheckoutPageFormStepDefs.java
@@ -1,4 +1,0 @@
-package com.example.Team1webshop.StepDefs;
-
-public class CheckoutPageFormStepDefs {
-}
