+++ src/test/java/com/example/Team1webshop/StepDefsTests/CheckoutPageStepDefs.java
@@ -1,4 +1,4 @@
-package com.example.Team1webshop.StepDefs;
+package com.example.Team1webshop.StepDefsTests;
 
 import io.cucumber.java.Before;
 import io.cucumber.java.en.Given;
